Resolve runtime warning of the "weight" property
Turns out the runtime warning was coming from the call to set the
"weight" property for the boldrenderer tree view cell renderer.

Removed the int cast for the other two "weight" property settings as it
is not needed as an enum is always an int type.  Duh.

+++ src/gui_settingsdialog.c
@@ -1161,16 +1161,12 @@ on_settings_activate(GtkMenuItem    *menuitem,
                                        &preferences.promptcolor,
                                        NULL);
             gui->calltagname = g_strdup_printf("%d", rand());
-
-            /* Cast PANGO_WEIGHT_BOLD to int to quell runtime warning:
-             * value "<integer>" of type 'gint' is invalid or out of range for property 'weight' of type 'gint'
-             */
             gtk_text_buffer_create_tag(buffer,
                                        gui->calltagname,
                                        "foreground-rgba",
                                        &preferences.promptcolor,
                                        "weight",
-                                       (int)PANGO_WEIGHT_BOLD,
+                                       PANGO_WEIGHT_BOLD,
                                        NULL);
             preferences.promptcolor = color;
         }
